Updated testcases
